refactor: 이미지 파일 webp로 변환
