+++ src/Pages/HomePage/index.tsx
@@ -2,9 +2,9 @@ import styled from '@emotion/styled';
 import { theme } from '@/styles/theme';
 import NavigationBar from '@/Shared/components/NavigationBar';
 import { Container } from '@/Shared/components/Container';
-import CharacterMain from '@/assets/HomeImg/character.png';
-import QuizIcon from '@/assets/HomeImg/quiz.png';
-import NewsIcon from '@/assets/HomeImg/news.png';
+import CharacterMain from '@/assets/HomeImg/character.webp';
+import QuizIcon from '@/assets/HomeImg/quiz.webp';
+import NewsIcon from '@/assets/HomeImg/news.webp';
 import { useNavigate } from 'react-router-dom';
 import { Header } from '@/Shared/components/Header';
 import { StatusActionBar } from '@/Shared/components/StatusActionBar';
