+++ src/Pages/TierPage/index.tsx
@@ -4,12 +4,12 @@ import { Header } from '@/Shared/components/Header';
 import { theme } from '@/styles/theme';
 import { useMemo } from 'react';
 
-import IconBoss from '@/assets/TierImg/tier_boss.png';
-import IconDividend from '@/assets/TierImg/tier_dividend.png';
-import IconBlueChip from '@/assets/TierImg/tier_bluechip.png';
-import IconGrowth from '@/assets/TierImg/tier_growth.png';
-import IconSeed from '@/assets/TierImg/tier_seed.png';
-import CharacterMain from '@/assets/HomeImg/character.png';
+import IconBoss from '@/assets/TierImg/tier_boss.webp';
+import IconDividend from '@/assets/TierImg/tier_dividend.webp';
+import IconBlueChip from '@/assets/TierImg/tier_bluechip.webp';
+import IconGrowth from '@/assets/TierImg/tier_growth.webp';
+import IconSeed from '@/assets/TierImg/tier_seed.webp';
+import CharacterMain from '@/assets/HomeImg/character.webp';
 
 type Tier = {
   grade: string;
